requirements.txt
ccxt>=4.5.0
pyyaml==6.0
aiohttp>=3.10.11
asyncio-mqtt==0.16.0
websockets==12.0
python-dotenv==1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
aiohttp>=3.8.0
websockets>=11.0.0
asynctest>=0.13.0
hyperliquid-python-sdk>=0.20.1
# lighter-api-python==1.0.0  # 如果Lighter有官方SDK
# paradex-python==1.0.0      # 如果Paradex有官方SDK